requirements.txt
python-telegram-bot==4.0.3
lxml==3.3.3
urllib3==1.15.1
pyopenssl
ndg-httpsclient pyasn1
